requirements.txt
requests==2.24.0
urllib3==1.25.9
kivymd==0.104.1
docutils==0.16
Pillow==7.1.2
Kivy==1.11.1
dj-database-url==0.5.0
Django==3.0.7
gunicorn==20.0.4
mysqlclient==1.4.6
psycopg2==2.8.5
whitenoise==5.1.0
djangorestframework==3.11.0
djoser==2.0.3
